+++ DevDirManager/internal/scripts/postimport.ps1
@@ -24,3 +24,6 @@ $moduleRoot = Split-Path (Split-Path $PSScriptRoot)
 
 # Load License
 "$moduleRoot\internal\scripts\license.ps1"
+
+# Load System Configuration
+"$moduleRoot\internal\scripts\configuration.ps1"